Moved services to fan entity
